refactor: 아키텍처 변경
