+++ supabase/functions/api/supabase.yaml
@@ -1,0 +1,4 @@
+functions:
+  api:
+    handler:
+      src: functions/api/index.ts
