+++ supabase/functions/api/users/userEntity.ts
